requirements.txt
pytest
pytest-asyncio
pytest-cov
black
isort
flake8
mypy
fastapi
uvicorn
pydantic
pydantic-settings
httpx
pinecone
pinecone-client
asyncio
python-dotenv
supabase
python-multipart
tenacity
loguru
langchain
langchain-community
langchain-core
tiktoken
openai
anthropic
google-generativeai
openai-whisper
pymupdf
python-docx
markdown
pandas
openpyxl
xlrd
python-pptx
pillow
opencv-python
ffmpeg-python
pyyaml
